refactor: modal창 위치 조정
