+++ src/app/(beforeLogin)/@modal/(.)i/flow/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginModal from "@/app/(beforeLogin)/_component/LoginModal";
+import LoginModal from "@/(beforeLogin)/_component/LoginModal";
 
 export default function Page() {
   console.log("모달입니당");
