+++ src/app/(beforeLogin)/_component/LoginModal.tsx
@@ -31,10 +31,10 @@ export default function LoginModal() {
   };
 
   return (
-    <div className="flex justify-center items-center">
+    <div className="fixed inset-0 bg-black bg-opacity-70 flex justify-center items-center">
       <div className="absolute inset-0 opacity-50"></div>
       <div className="relative max-w-[80vw] min-w-[550px] h-[500px] bg-black rounded-lg">
-        <div>
+        <div className="flex">
           <button
             className="w-8 h-8 flex justify-center items-center"
             onClick={onClickClose}
@@ -48,6 +48,17 @@ export default function LoginModal() {
               <path d="M10.59 12L4.54 5.96l1.42-1.42L12 10.59l6.04-6.05 1.42 1.42L13.41 12l6.05 6.04-1.42 1.42L12 13.41l-6.04 6.05-1.42-1.42L10.59 12z" />
             </svg>
           </button>
+          <div className="ml-50">
+            <svg
+              viewBox="0 0 24 24"
+              aria-hidden="true"
+              className="w-8 h-8 fill-white r-4qtqp9 r-yyyyoo r-dnmrzs r-bnwqim r-lrvibr r-m6rgpd r-1nao33i r-lwhw9o r-cnnz9e"
+            >
+              <g>
+                <path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z" />
+              </g>
+            </svg>
+          </div>
         </div>
         <div className="flex flex-col justify-center items-center flex-grow space-y-6">
           <h1 className="text-3xl font-bold mb-12">X 가입하기</h1>
