+++ src/app/(beforeLogin)/i/flow/single_sign_on/page.tsx
@@ -1,4 +1,4 @@
-import SingleSignOn from "@/app/(beforeLogin)/_component/SingleSignOn";
+import SingleSignOn from "@/(beforeLogin)/_component/SingleSignOn";
 
 export default function Page() {
   return <SingleSignOn />;
