+++ src/app/(beforeLogin)/page.tsx
@@ -1,4 +1,4 @@
-import Main from "@/app/(beforeLogin)/_component/Main";
+import Main from "@/(beforeLogin)/_component/Main";
 
 export default function Home() {
   return <Main />;
